requirements.txt
# Core dependencies
python-dotenv>=1.0.0
langchain>=0.1.0
openai>=1.0.0
jupyter>=1.0.0
notebook>=7.0.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Vector database and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Utilities
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0
argparse>=1.4.0 